+++ lib/models/exceptions/exceptions.dart
@@ -1,0 +1,2 @@
+export 'bad_type_exception.dart';
+export 'missing_field_exception.dart';