+++ lib/utils/maths/maths.dart
@@ -1,4 +1,2 @@
-library geoff.utils.maths;
-
 export 'random_utils.dart';
 export 'number_utils.dart';