+++ lib/utils/networking/networking.dart
@@ -1,6 +1,4 @@
-library geoff.utils.networking;
-
-export 'gprc_utils.dart';
-export 'may_rest.dart';
-export 'status_code.dart';
+export './gprc_utils.dart';
+export './may_rest.dart';
+export './status_code.dart';
 export './auth/auth.dart';