+++ lib/utils/system/sensors/sensors.dart
@@ -1,1 +1,1 @@
-library geoff.utils.system.sensors;
+export "./orientation_utils.dart";