+++ lib/utils/utils.dart
@@ -1,7 +1,7 @@
-library geoff.utils;
-
 export './networking/networking.dart';
 export './system/system.dart';
 export './maths/maths.dart';
 export './alarm.dart';
-export 'time/time.dart';
+export './functional/functional.dart';
+export './time/time.dart';
+export './string/string.dart';