+++ CHANGELOG.md
@@ -1,6 +1,17 @@
 # **Changelog**
 
 ## **Version 0.2**
+### **0.2.5**
+- Changed a bunch of classes with only `static` methods to be `abstract` and added private constructors to prevent extending
+- Added the `functional` library, which contains a few funtional programming methods.
+- Added `time` library with some utc helpers.
+
+Finally adding stuff to the `models` package this update!
+- Added `GeoffUser` and `KeycloakUser`
+- Added `ModelUtils`
+- Added `BadTypeException` and `MissingFieldException`
+- Added the `RichTextBlock` model, which can be used to output rich text with html tags! Currently no editor for it, maybe down the line I'll do that
+
 ### **0.2.4**
 - Re-wrote `MayRest` to reduce code duplication
 - `MayRest` calls can now be manually given a token